Modify validation and auth in orders API endpoints. Include ticket endpoint in API. Fix docker-compose.yml.

+++ app-service/src/main/java/es/mqm/webapp/model/Order.java
@@ -216,7 +216,7 @@ public class Order {
         return totalPrice;
     }
 
-    public void setTotalPrice(int totalPrice) {
+    public void setTotalPrice(double totalPrice) {
         this.totalPrice = totalPrice;
     }
 }
